Basics
Program is fully functional. It is able to find the url of an image,
download it, make a folder to store it in, then set that image as the
background. Optimization is still needed as well as a GUI. Resolution is
determined but no yet implemented, not sure if needed.